// src/app/models/mature.profile.chart.model.ts
import { Section1Model } from './sections/section1.model';
import { Section2Model } from './sections/section2.model';
import { Section3Model } from './sections/section3.model';
import { Section4Model } from './sections/section4.model';
import { Section5Model } from './sections/section5.model';
import { Section6Model } from './sections/section6.model';
import { Section7Model } from './sections/section7.model';

export class MatureProfileChartModel {
    section1: Section1Model;
    section2: Section2Model;
    section3: Section3Model;
    section4: Section4Model;
    section5: Section5Model;
    section6: Section6Model;
    section7: Section7Model;
}